leagueGames/requirements.txt
python-dotenv==1.0.1
pymongo===4.7.1
pandas==2.2.2
requests==2.31.0
schedule
flask